requirement.txt
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-gis==0.18
GDAL==3.4.1
numpy==1.22.3
psycopg2==2.9.3
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
